% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_directory_metadata.R
\name{assert_databicks_path}
\alias{assert_databicks_path}
\title{Validate the path}
\usage{
assert_databicks_path(path)
}
\arguments{
\item{path}{Path to be validated}
}
\value{
Invisible path
}
\description{
The assert_path function validates the path for file system operations.
}
