% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_api.R
\name{assert_databricks_path}
\alias{assert_databricks_path}
\title{Validate the path}
\usage{
assert_databricks_path(path)
}
\arguments{
\item{path}{Path to be validated}
}
\value{
Invisible path
}
\description{
The assert_path function validates the path for file system operations.
}
