% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_methods_utils.R
\name{cnt_create_directory}
\alias{cnt_create_directory}
\title{Create a directory}
\usage{
cnt_create_directory(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The \link{ConnectorDatabricksVolume} object}

\item{name}{The name of the directory to create}

\item{...}{Additional parameters to pass to the files_create_directory method}
}
\value{
Information about newly created directory
}
\description{
Create a directory
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# databricks_volume is a ConnectorDatabricksVolume object
databricks_volume \%>\%
  cnt_create_directory("folder")
# This function is used by the method create_directory
databricks_volume$create_directory("folder")
\dontshow{\}) # examplesIf}
}
