% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R
\docType{import}
\name{cnt_disconnect}
\alias{cnt_disconnect}
\title{Objects exported from other packages}
\value{
\link{invisible} connector_object.
}
\description{
Generic implementing of how to disconnect from the relevant connections.
Mostly relevant for DBI connectors.

\itemize{
\item \link[connector]{connector_dbi}: Uses \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}} to create a table reference to close a DBI connection.
}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{connector}{\code{\link[connector]{cnt_disconnect}}}
}}

