% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/dbi_methods.R,
%   R/volume_methods.R
\name{cnt_list_content}
\alias{cnt_list_content}
\alias{cnt_list_content.connector_databricks_dbi}
\alias{cnt_list_content.ConnectorDatabricksVolume}
\title{List available content from the connector}
\usage{
cnt_list_content(connector_object, ...)

\method{cnt_list_content}{connector_databricks_dbi}(connector_object, ...)

\method{cnt_list_content}{ConnectorDatabricksVolume}(connector_object, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
A \link{character} vector of content names
}
\description{
Addition list content methods for databricks connectors implemented for \code{\link[connector:cnt_list_content]{connector::cnt_list_content()}}:

\itemize{
\item \link{connector_databricks_dbi}: Reuses the \code{\link[connector:cnt_list_content]{connector::cnt_list_content()}} method for \link[connector:connector_dbi]{connector::connector_dbi},
but always sets the \code{catalog} and \code{schema} as defined in when initializing the connector.
}

\itemize{
\item \link{ConnectorDatabricksVolume}: Reuses the \code{\link[connector:cnt_list_content]{connector::cnt_list_content()}}
method for \link{ConnectorDatabricksVolume}, but always sets the
\code{catalog} and \code{schema} as defined in when initializing the connector.
}
}
