% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/dbi_methods.R
\name{cnt_read}
\alias{cnt_read}
\alias{cnt_read.connector_databricks_dbi}
\title{Read content from the connector}
\usage{
cnt_read(connector_object, name, ...)

\method{cnt_read}{connector_databricks_dbi}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
R object with the content. For rectangular data a \link{data.frame}.
}
\description{
Addition read methods for databricks connectors implemented for \code{\link[connector:cnt_read]{connector::cnt_read()}}:

\itemize{
\item \link{connector_databricks_dbi}: Reuses the \code{\link[connector:cnt_write]{connector::cnt_write()}} method for \link[connector:connector_dbi]{connector::connector_dbi},
but always sets the \code{catalog} and \code{schema} as defined in when initializing the connector.
}
}
