% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/dbi_methods.R
\name{cnt_remove}
\alias{cnt_remove}
\alias{cnt_remove.connector_databricks_dbi}
\title{Remove content from the connector}
\usage{
cnt_remove(connector_object, name, ...)

\method{cnt_remove}{connector_databricks_dbi}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Addition remove methods for databricks connectors implemented for \code{\link[connector:cnt_remove]{connector::cnt_remove()}}:

\itemize{
\item \link{connector_databricks_dbi}: Reuses the \code{\link[connector:cnt_remove]{connector::cnt_remove()}} method for \link[connector:connector_dbi]{connector::connector_dbi},
but always sets the \code{catalog} and \code{schema} as defined in when initializing the connector.
}
}
