% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/dbi_methods.R
\name{cnt_tbl}
\alias{cnt_tbl}
\alias{cnt_tbl.connector_databricks_dbi}
\title{Use dplyr verbs to interact with the remote database table}
\usage{
cnt_tbl(connector_object, name, ...)

\method{cnt_tbl}{connector_databricks_dbi}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
A \link[dplyr:tbl]{dplyr::tbl} object.
}
\description{
Addition tbl methods for databricks connectors implemented for \code{\link[connector:cnt_tbl]{connector::cnt_tbl()}}:

\itemize{
\item \link{connector_databricks_dbi}: Reuses the \code{\link[connector:cnt_tbl]{connector::cnt_tbl()}} method for \link[connector:connector_dbi]{connector::connector_dbi},
but always sets the \code{catalog} and \code{schema} as defined in when initializing the connector.
}
}
