% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector-databricks-options.R
\name{connector-databricks-options-params}
\alias{connector-databricks-options-params}
\title{Internal parameters for reuse in functions}
\arguments{
\item{overwrite}{Overwrite existing content if it exists in the connector?. Default: \code{FALSE}.}

\item{verbosity_level}{Verbosity level for functions in connector. See \link[zephyr:verbosity_level]{zephyr::verbosity_level} for details.. Default: \code{"verbose"}.}
}
\description{
Internal parameters for reuse in functions
}
\details{
See \link{connector-options-databricks} for more information.
}
\keyword{internal}
