% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.databricks-package.R
\docType{package}
\name{connector.databricks-package}
\alias{connector.databricks}
\alias{connector.databricks-package}
\title{connector.databricks: Expansion of the connector package for enabling connections to databricks}
\description{
Expands the connector package to handle databricks. This includes reading and writing of files to volumes as well as connections to sparklyr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jubilant-adventure-yr95p3g.pages.github.io/}
}

}
\author{
\strong{Maintainer}: Steffen Falgreen Larsen \email{sffl@novonordisk.com}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
