% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks.R
\name{connector_databricks}
\alias{connector_databricks}
\title{Create databricks connector object}
\usage{
connector_databricks(
  host,
  token,
  cluster_id,
  catalog,
  schema,
  ...,
  extra_class = NULL
)
}
\arguments{
\item{host}{Your Workspace Instance URL.}

\item{token}{Your Personal Authentication Token, it will be retrieved
by \link{get_token}}

\item{cluster_id}{Cluster id. More about it can be find
\href{https://docs.databricks.com/en/workspace/workspace-details.html#cluster-url-and-id}{here}.}

\item{catalog}{Catalog name. More about it can be find
\href{https://docs.databricks.com/en/catalogs/index.html}{here}.}

\item{schema}{Schema name. More about it can be find
\href{https://docs.databricks.com/en/schemas/index.html}{here}.}

\item{...}{Additional parameters to pass to the \link{Connector_databricks} object}

\item{extra_class}{\link{character} Extra class added to the object. See details.}
}
\value{
A new \link{Connector_databricks} object
}
\description{
Create a new databricks connector object. See \link{Connector_databricks} for details.
}
\details{
The \code{extra_class} parameter allows you to create a subclass of the \code{Connector_databricks} object.
This can be useful if you want to create a custom connection object for easier dispatch of new s3 methods,
while still inheriting the methods from the \code{Connector_databricks} object.
}
\examples{
\dontshow{if (not_on_ci()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Connect

db_con <- connector_databricks(
  host = Sys.getenv("DATABRICKS_HOST"),
  token = Sys.getenv("DATABRICKS_TOKEN"),
  cluster_id = "0216-121054-v8tdvp00",
  catalog = "samples",
  schema = "nyctaxi"
)
# List content
db_con$list_content()
# Write a file
db_con$write(iris, "iris.csv")
# Read a file
db_con$read("iris.csv")
# Remove a file or directory
db_con$remove("iris.csv", confirm = FALSE)
\dontshow{\}) # examplesIf}
}
