% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{connector_databricks_dbi}
\alias{connector_databricks_dbi}
\title{Connector for connecting to Databricks using DBI}
\description{
Extension of the \link[connector:connector_dbi]{connector::connector_dbi} making it easier to connect to, and work with tables in Databricks.
}
\details{
All methods for \link{connector_databricks_dbi} object are working from the catalog
and schema provided when initializing the connection.
This means you only need to provide the table name when using the built in methods.
If you want to access tables outside of the chosen schema, you can either retrieve
the connection with \code{connector_databricks_dbi$conn} or create a new connector.

When creating the connections to Databricks you either need to provide the sqlpath to
the Databricks cluster or the SQL warehouse you want to connect to.
Authentication to databricks is handed by the \code{odbc::databricks()} driver and supports
general use of personal access tokens and credentials through Posit Workbench.
See also \code{\link[odbc:databricks]{odbc::databricks()}} On more information on how the connection to Databricks is established.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Establish connection to your cluster

con_databricks <- connector_databricks_dbi$new(
httpPath = "path-to-cluster",
  catalog = "my_catalog",
  schema = "my_schema"
)

# List tables in my_schema

con_databricks$list_content()

# Read and write tables

con_databricks$write(mtcars, "my_mtcars_table")

con_databricks$read("my_mtcars_table")

# Use dplyr::tbl

con_databricks$tbl("my_mtcars_table")

# Remove table

con_databricks$remove("my_mtcars_table")

# Disconnect

con_databricks$disconnect()
\dontshow{\}) # examplesIf}
}
\section{Super classes}{
\code{\link[connector:connector]{connector::connector}} -> \code{\link[connector:connector_dbi]{connector::connector_dbi}} -> \code{connector_databricks_dbi}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{conn}}{The DBI connection object of the connector}

\item{\code{catalog}}{The catalog used in the connector}

\item{\code{schema}}{The schema used in the connector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connector_databricks_dbi-new}{\code{connector_databricks_dbi$new()}}
\item \href{#method-connector_databricks_dbi-clone}{\code{connector_databricks_dbi$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_list_content"><a href='../../connector/html/connector.html#method-connector-cnt_list_content'><code>connector::connector$cnt_list_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_read"><a href='../../connector/html/connector.html#method-connector-cnt_read'><code>connector::connector$cnt_read()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_remove"><a href='../../connector/html/connector.html#method-connector-cnt_remove'><code>connector::connector$cnt_remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_write"><a href='../../connector/html/connector.html#method-connector-cnt_write'><code>connector::connector$cnt_write()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="print"><a href='../../connector/html/connector.html#method-connector-print'><code>connector::connector$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector_dbi" data-id="cnt_disconnect"><a href='../../connector/html/connector_dbi.html#method-connector_dbi-cnt_disconnect'><code>connector::connector_dbi$cnt_disconnect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector_dbi" data-id="cnt_tbl"><a href='../../connector/html/connector_dbi.html#method-connector_dbi-cnt_tbl'><code>connector::connector_dbi$cnt_tbl()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_databricks_dbi-new"></a>}}
\if{latex}{\out{\hypertarget{method-connector_databricks_dbi-new}{}}}
\subsection{Method \code{new()}}{
Initialize the connection to Databricks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_databricks_dbi$new(http_path, catalog, schema)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{http_path}}{\link{character} The path to the Databricks cluster or SQL warehouse you want to connect to}

\item{\code{catalog}}{\link{character} The catalog to use}

\item{\code{schema}}{\link{character} The schema to use}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{connector_databricks_dbi} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connector_databricks_dbi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-connector_databricks_dbi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{connector_databricks_dbi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
