% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks.R
\name{Connector_databricks_object}
\alias{Connector_databricks_object}
\alias{Connector_databricks}
\title{Connector Object for databricks}
\description{
This object is used to interact with databricks spark cluster,
adding the ability to list, read, write, download, upload, create tables
and remove files.
}
\details{
About the token, you can retrieve it by following the guideline in your
enterprise.
}
\examples{
\dontshow{if (not_on_ci()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Connect to databricks
cluster <- Connector_databricks$new(
  host = Sys.getenv("DATABRICKS_HOST"),
  token = Sys.getenv("DATABRICKS_TOKEN"),
  cluster_id = Sys.getenv("DATABRICKS_CLUSTER_ID"),
  catalog = catalog,
  schema = schema
)

my_drive$list_content()
\dontshow{\}) # examplesIf}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cluster_id}}{(\code{string}) Cluster ID}

\item{\code{catalog}}{(\code{string}) Catalog}

\item{\code{schema}}{(\code{string}) Schema}

\item{\code{sc}}{(\verb{Spark connection}) Spark Connection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Connector_databricks-new}{\code{Connector_databricks$new()}}
\item \href{#method-Connector_databricks-list_tables}{\code{Connector_databricks$list_tables()}}
\item \href{#method-Connector_databricks-read}{\code{Connector_databricks$read()}}
\item \href{#method-Connector_databricks-write}{\code{Connector_databricks$write()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks-new"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Connector_databricks class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks$new(
  host = Sys.getenv("DATABRICKS_HOST"),
  token = Sys.getenv("DATABRICKS_TOKEN"),
  cluster_id,
  catalog,
  schema,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{Your Workspace Instance URL.}

\item{\code{token}}{Your Personal Authentication Token, it will be retrieved
by \link{get_token}}

\item{\code{cluster_id}}{Cluster id. More about it can be find
\href{https://docs.databricks.com/en/workspace/workspace-details.html#cluster-url-and-id}{here}.}

\item{\code{catalog}}{Catalog name. More about it can be find
\href{https://docs.databricks.com/en/catalogs/index.html}{here}.}

\item{\code{schema}}{Schema name. More about it can be find
\href{https://docs.databricks.com/en/schemas/index.html}{here}.}

\item{\code{...}}{Additional parameters to pass to the \link{Connector_databricks} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{Connector_databricks} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks-list_tables"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks-list_tables}{}}}
\subsection{Method \code{list_tables()}}{
List tables of the schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks$list_tables(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters to pass to the cnt_list_content method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataframe with list of tables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks-read"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks-read}{}}}
\subsection{Method \code{read()}}{
Read the content of a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks$read(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file to read}

\item{\code{...}}{Additional parameters to pass to the cnt_read method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The content of the file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks-write"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks-write}{}}}
\subsection{Method \code{write()}}{
Write a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks$write(x, file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The content to write}

\item{\code{file}}{The name of the file to write}

\item{\code{...}}{Additional parameters to pass to the cnt_write method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The content of the file
,
#' @description Download a file
#' @param name The name of the file to download
#' @param ... Additional parameters to pass to the cnt_download_content method
#' @return The file downloaded
download = function(name, ...) {
self \%>\%
cnt_download_content(name, ...)
},
#' @description Upload a file
#' @param name The name of the file to upload
#' @param ... Additional parameters to pass to the cnt_upload_content method
#' @return The file uploaded
upload = function(name, ...) {
self \%>\%
cnt_upload_content(name, ...)
},
#' @description Get the connection
#' @return The connection
get_conn = function() {
private$sc
},
#' @description Create a directory
#' @param name The name of the directory to create
#' @param ... Additional parameters to pass to the cnt_create_directory method
#' @return The directory created
create_directory = function(name, ...) {
self \%>\%
cnt_create_directory(name, ...)
},
#' @description Remove a file or a directory
#' @param name The name of the file or directory to remove
#' @param ... Additional parameters to pass to the cnt_remove method
remove = function(name, ...) {
self \%>\%
cnt_remove(name, ...)
}
}
}
}
