% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{connector_databricks_sql}
\alias{connector_databricks_sql}
\title{Create \code{ConnectorDatabricksSQL} connector
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
connector_databricks_sql(
  warehouse_id = Sys.getenv("DATABRICKS_WAREHOUSE_ID"),
  catalog,
  schema,
  staging_volume = NULL,
  ...,
  extra_class = NULL
)
}
\arguments{
\item{warehouse_id}{\link{character} The ID of the Databricks SQL warehouse
you want to connect to. Defaults to \code{DATABRICKS_WAREHOUSE_ID} environment
variable.}

\item{catalog}{\link{character} The catalog to use}

\item{schema}{\link{character} The schema to use}

\item{staging_volume}{\link{character} Optional volume path for large dataset
staging. Recommended way for better performances.}

\item{...}{Additional parameters sent to \code{brickster::DatabricksSQL()}
driver.}

\item{extra_class}{\link{character} Extra class to assign to the new connector}
}
\value{
A \link{ConnectorDatabricksSQL} object
}
\description{
Initializes the connector for SQL warehouse type of storage.
See \link{ConnectorDatabricksSQL} for details.

Initialize the connection to Databricks
}
\details{
The \code{extra_class} parameter allows you to create a subclass of the
\code{ConnectorDatabricksSQL} object. This can be useful if you want to create
a custom connection object for easier dispatch of new s3 methods, while still
inheriting the methods from the \code{ConnectorDatabricksSQL} object.
}
\examples{
\dontrun{
# Establish connection to your SQL warehouse

con_databricks <- connector_databricks_sql(
  warehouse_id = "your-warehouse-id",
  catalog = "my_catalog",
  schema = "my_schema"
)

# List tables in my_schema

con_databricks$list_content()

# Read and write tables

con_databricks$write(mtcars, "my_mtcars_table")

con_databricks$read("my_mtcars_table")

# Use dplyr::tbl

con_databricks$tbl("my_mtcars_table")

# Remove table

con_databricks$remove("my_mtcars_table")

# Disconnect

con_databricks$disconnect()
}
}
