% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks_volumes.R
\name{connector_databricks_volumes}
\alias{connector_databricks_volumes}
\title{Create databricks volumes connector}
\usage{
connector_databricks_volumes(path, ..., extra_class = NULL)
}
\arguments{
\item{path}{Path to the file system}

\item{...}{Additional arguments passed to \link{Connector_databricks_volumes}}

\item{extra_class}{\link{character} Extra class added to the object. See details.}
}
\value{
A new \link{Connector_databricks_volumes} object
}
\description{
Create a new databricks volumes connector object. See \link{Connector_databricks_volumes} for details.
}
\details{
The \code{extra_class} parameter allows you to create a subclass of the \code{Connector_databricks_volumes} object.
This can be useful if you want to create a custom connection object for easier dispatch of new s3 methods,
while still inheriting the methods from the \code{Connector_databricks_volumes} object.
}
\examples{
# Connect to a file system

databrics_volume <- "/Volumes/amace_cdr_bronze_dev/nn9536_4373_adam/tester"
db <- connector_databricks_volumes(databrics_volume)

db

# Create subclass connection

db_subclass <- connector_databricks_volumes(databrics_volume, extra_class = "subclass")

db_subclass
class(db_subclass)

}
