% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks_volumes.R
\name{Connector_databricks_volumes_object}
\alias{Connector_databricks_volumes_object}
\alias{Connector_databricks_volumes}
\title{Class Connector_databricks_volumes}
\description{
The connector_databricks_volumes class is a file system connector
for accessing and manipulating files in a local file system.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Connector_databricks_volumes-new}{\code{Connector_databricks_volumes$new()}}
\item \href{#method-Connector_databricks_volumes-list_content}{\code{Connector_databricks_volumes$list_content()}}
\item \href{#method-Connector_databricks_volumes-construct_path}{\code{Connector_databricks_volumes$construct_path()}}
\item \href{#method-Connector_databricks_volumes-read}{\code{Connector_databricks_volumes$read()}}
\item \href{#method-Connector_databricks_volumes-write}{\code{Connector_databricks_volumes$write()}}
\item \href{#method-Connector_databricks_volumes-create_directory}{\code{Connector_databricks_volumes$create_directory()}}
\item \href{#method-Connector_databricks_volumes-remove}{\code{Connector_databricks_volumes$remove()}}
\item \href{#method-Connector_databricks_volumes-remove_directory}{\code{Connector_databricks_volumes$remove_directory()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks_volumes-new"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks_volumes-new}{}}}
\subsection{Method \code{new()}}{
Initializes the connector_databricks_volumes class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks_volumes$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the file system}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks_volumes-list_content"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks_volumes-list_content}{}}}
\subsection{Method \code{list_content()}}{
Returns the list of files in the specified path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks_volumes$list_content(
  ...,
  full.names = FALSE,
  client = DatabricksClient()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{used to construct nested paths}

\item{\code{full.names}}{a logical value. If TRUE, the directory path is
prepended to the file names to give a relative file path.
If FALSE, the file names (rather than paths) are returned.}

\item{\code{client}}{the Databricks client}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks_volumes-construct_path"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks_volumes-construct_path}{}}}
\subsection{Method \code{construct_path()}}{
Constructs a complete path by combining the specified access path with the provided elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks_volumes$construct_path(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Elements to construct the path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks_volumes-read"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks_volumes-read}{}}}
\subsection{Method \code{read()}}{
Reads the content of the specified file using the private access path and additional options
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks_volumes$read(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the file to read}

\item{\code{...}}{Other parameters to pass to the read_file function (depends on the extension of a file)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks_volumes-write"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks_volumes-write}{}}}
\subsection{Method \code{write()}}{
Writes the specified content to the specified file using the
private access path and additional options
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks_volumes$write(x, file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Content to write to the file}

\item{\code{file}}{File name}

\item{\code{...}}{Other parameters to pass to the write_file function (depends on the extension of a file)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks_volumes-create_directory"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks_volumes-create_directory}{}}}
\subsection{Method \code{create_directory()}}{
Create directory on databricks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks_volumes$create_directory(dir, client = DatabricksClient())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{directory name}

\item{\code{client}}{a databricks client made using DatabricksClient}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks_volumes-remove"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks_volumes-remove}{}}}
\subsection{Method \code{remove()}}{
Remove the specified file by given name with extension
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks_volumes$remove(file, client = DatabricksClient())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{File name}

\item{\code{client}}{a databricks client made using DatabricksClient}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_databricks_volumes-remove_directory"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_databricks_volumes-remove_directory}{}}}
\subsection{Method \code{remove_directory()}}{
Remove the specified directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_databricks_volumes$remove_directory(dir, client = DatabricksClient())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{directory name}

\item{\code{client}}{a databricks client made using DatabricksClient}
}
\if{html}{\out{</div>}}
}
}
}
