% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cnt_list_content.Connector_databricks}
\alias{cnt_list_content.Connector_databricks}
\alias{cnt_read.Connector_databricks}
\alias{cnt_write.Connector_databricks}
\title{Methods for the databricks class}
\usage{
\method{cnt_list_content}{Connector_databricks}(connector_object)

\method{cnt_read}{Connector_databricks}(connector_object, name, ...)

\method{cnt_write}{Connector_databricks}(connector_object, x, file, ...)
}
\arguments{
\item{connector_object}{The databricks object}

\item{name}{The name of the file to read}

\item{...}{Additional parameters to pass to the write_microsoft_file function}

\item{x}{The content to write}

\item{file}{The name of the file to write}
}
\description{
Methods for the databricks class
}
