% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_methods_utils.R
\name{create_directory_cnt}
\alias{create_directory_cnt}
\title{Create a directory}
\usage{
create_directory_cnt(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The \link{ConnectorDatabricksVolume} object}

\item{name}{The name of the directory to create}

\item{...}{Additional parameters to pass to the files_create_directory method}
}
\value{
New \link{ConnectorDatabricksVolume} object of a newly built directory
}
\description{
Create a directory
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# databricks_volume is a ConnectorDatabricksVolume object
new_connector <- databricks_volume \%>\%
  create_directory_cnt("folder")
# This function is used by the method create_directory
new_connector <- databricks_volume$create_directory("folder")
\dontshow{\}) # examplesIf}
}
