% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabricksClient.R
\name{DatabricksClient}
\alias{DatabricksClient}
\title{DatabricksClient is a constructor for class that performs any operations with
Databricks REST API and handle a subset of Unified Client Authentication.}
\usage{
DatabricksClient(profile = NULL, host = NULL, token = NULL, config_file = NULL)
}
\arguments{
\item{profile}{configuration profile from ~/.databrickscfg. Defaults to DEFAULT}

\item{host}{URL of Databricks Workspace}

\item{token}{Personal Access Token}

\item{config_file}{path to Databricks CLI configuration file. Defaults to ~/.databrickscfg}
}
\description{
DatabricksClient is a constructor for class that performs any operations with
Databricks REST API and handle a subset of Unified Client Authentication.
}
\examples{
\dontrun{
# In order to connect to databricks on environments where configurations are
# available via the environment variable DATABRICKS_CONFIG_FILE or located
# at ~/.databrickscfg - simply write
db_client <- DatabricksClient()
# To check if connection is established
open_connection <- db_client$debug_string() != ""

if (open_connection) {
  db_client$do("GET", "/api/2.1/unity-catalog/catalogs")
}
}
}
