% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R
\name{download_directory_cnt}
\alias{download_directory_cnt}
\title{Download a directory}
\usage{
download_directory_cnt(connector_object, name, dir = name, ...)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{name}{\link{character} The name of the directory to download}

\item{dir}{\link{character} Path to the directory to download to}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Addition list content methods for databricks connectors implemented for
\code{\link[connector:download_directory_cnt]{connector::download_directory_cnt()}}:
}
