% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Downloads a file of up to 5 GiB. The file contents are the response body.
This is a standard HTTP file download, not a JSON RPC.}
\usage{
download_file(client, file_path, local_path = NULL)
}
\arguments{
\item{client}{Required. Instance of DatabricksClient()}

\item{file_path}{Required. The absolute path of the file.}

\item{local_path}{local path for the file.
if \code{NULL} the response from databricks is returned.}
}
\value{
The repsonse from databricks.
If \code{!is.null(local_path)} the response is returned invisible.
}
\description{
Downloads a file of up to 5 GiB. The file contents are the response body.
This is a standard HTTP file download, not a JSON RPC.
}
