% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_api.R
\name{files_create_directory}
\alias{files_create_directory}
\title{Create a directory.}
\usage{
files_create_directory(directory_path, client = DatabricksClient())
}
\arguments{
\item{directory_path}{The absolute path of a directory.}

\item{client}{Instance of DatabricksClient()}
}
\description{
Creates an empty directory. If necessary, also creates any parent directories
of the new, empty directory (like the shell command \code{mkdir -p}). If called on
an existing directory, returns a success response; this method is idempotent
(it will succeed if the directory already exists).
}
