% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_api.R
\name{files_delete_directory}
\alias{files_delete_directory}
\title{Delete a directory.}
\usage{
files_delete_directory(directory_path, client = DatabricksClient())
}
\arguments{
\item{directory_path}{Required. The absolute path of a directory.}

\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Deletes an empty directory.
}
\details{
To delete a non-empty directory, first delete all of its contents. This can
be done by listing the directory contents and deleting each file and
subdirectory recursively.
}
