% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_api.R
\name{files_delete_file}
\alias{files_delete_file}
\title{Delete a file.}
\usage{
files_delete_file(file_path, client = DatabricksClient())
}
\arguments{
\item{file_path}{Required. The absolute path of the file.}

\item{client}{Required. Instance of DatabricksClient()}
}
\description{
Deletes a file. If the request is successful, there is no response body.
}
