% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_api.R
\name{files_upload_file}
\alias{files_upload_file}
\title{Uploads a file of up to 5 GiB. The file contents should be sent as the
request body as raw bytes (an octet stream); do not encode or otherwise
modify the bytes before sending. The contents of the resulting file will be
exactly the bytes sent in the request body. If the request is successful,
there is no response body.}
\usage{
files_upload_file(
  file_path,
  contents,
  overwrite = TRUE,
  client = DatabricksClient()
)
}
\arguments{
\item{file_path}{The absolute path of the file.}

\item{contents}{File content}

\item{overwrite}{If true, an existing file will be overwritten.}

\item{client}{Instance of DatabricksClient()}
}
\description{
Uploads a file of up to 5 GiB. The file contents should be sent as the
request body as raw bytes (an octet stream); do not encode or otherwise
modify the bytes before sending. The contents of the resulting file will be
exactly the bytes sent in the request body. If the request is successful,
there is no response body.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# In order to connect to databricks on environments where configurations are
# available via the environment variable DATABRICKS_CONFIG_FILE or located
# at ~/.databrickscfg - simply write
db_client <- DatabricksClient()

# To check if connection is established
open_connection <- db_client$debug_string() != ""

if (open_connection) {
  tempfile_orig <- tempfile(fileext = ".txt")
  writeLines(letters, tempfile_orig)

  databrics_volume <- "/Volumes/amace_cdr_bronze_dev/nn9536_4373_adam/tester"
  databricks_file <- file.path(databrics_volume, basename(tempfile_orig))

  connector.databricks:::upload_file(db_client, databricks_file, contents = tempfile_orig)
}
\dontshow{\}) # examplesIf}
}
