% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_file_databricks.R
\name{find_file_databricks}
\alias{find_file_databricks}
\title{Find File}
\usage{
find_file_databricks(name, root, client = DatabricksClient())
}
\arguments{
\item{name}{Name of a file}

\item{root}{Path to the root folder}

\item{client}{DatabricksClient}
}
\value{
A full name path to the file or a error if multiples files or 0.
}
\description{
Find File
}
