% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/dbi_methods.R,
%   R/volume_methods.R
\name{list_content_cnt}
\alias{list_content_cnt}
\alias{list_content_cnt.connector_databricks_dbi}
\alias{list_content_cnt.ConnectorDatabricksVolume}
\title{List available content from the connector}
\usage{
list_content_cnt(connector_object, ...)

\method{list_content_cnt}{connector_databricks_dbi}(connector_object, ...)

\method{list_content_cnt}{ConnectorDatabricksVolume}(connector_object, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
A \link{character} vector of content names
}
\description{
Addition list content methods for databricks connectors implemented for \code{\link[connector:list_content_cnt]{connector::list_content_cnt()}}:

\itemize{
\item \link{connector_databricks_dbi}: Reuses the \code{\link[connector:list_content_cnt]{connector::list_content_cnt()}} method for \link[connector:connector_dbi]{connector::connector_dbi},
but always sets the \code{catalog} and \code{schema} as defined in when initializing the connector.
}

\itemize{
\item \link{ConnectorDatabricksVolume}: Reuses the \code{\link[connector:list_content_cnt]{connector::list_content_cnt()}}
method for \link{ConnectorDatabricksVolume}, but always sets the
\code{catalog} and \code{schema} as defined in when initializing the connector.
}
}
