% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{list_content_tags}
\alias{list_content_tags}
\title{List Databricks tables in a catalog based on tag values}
\usage{
list_content_tags(connector_object, tags)
}
\arguments{
\item{connector_object}{\link{ConnectorDatabricksTable} object for interacting
with Databricks}

\item{tags}{String containing tag names and tag values in SQL query format,
e.g. \verb{(tag_name1 = 'tag_value1' OR tag_name2 = 'tag_value2')}.
More info \href{https://docs.databricks.com/aws/en/database-objects/tags}{here}}
}
\value{
None
}
\description{
This function is used to list tables in a Databricks catalog based on a tag.
}
\examples{
\dontrun{
connector_object |>
    list_content_tags(tags = ("tag_value = 'value1'")))
}
}
