% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumes_api.R
\name{list_databricks_volumes}
\alias{list_databricks_volumes}
\title{List Databricks volumes}
\usage{
list_databricks_volumes(
  catalog_name,
  schema_name,
  max_results = 100,
  page_token = NULL,
  include_browse = FALSE,
  client = DatabricksClient()
)
}
\arguments{
\item{catalog_name}{The name of the catalog where the schema and the volume
are.}

\item{schema_name}{The name of the schema where the volume will be created.}

\item{max_results}{Maximum number of volumes to return (page length).
More info on the parameter can be found here:
https://docs.databricks.com/api/workspace/volumes/list#max_results}

\item{page_token}{Opaque token returned by a previous request. It must be
included in the request to retrieve the next page of results (pagination).}

\item{include_browse}{Whether to include volumes in the response for which
the principal can only access selective metadata for}

\item{client}{Instance of DatabricksClient().}
}
\value{
Data.frame with information about available volumes.
}
\description{
Returns list of Databricks volume.
}
\details{
More details can be found here:
https://docs.databricks.com/api/workspace/volumes/list
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# In order to connect to databricks on environments where configurations are
# available via the environment variable DATABRICKS_CONFIG_FILE or located
# at ~/.databrickscfg - simply write
db_client <- DatabricksClient()
# To check if connection is established
open_connection <- db_client$debug_string() != ""

if (open_connection) {
  connector.databricks::list_databricks_volumes(db_client,
    catalog_name = "amace_cdr_bronze_dev",
    schema_name = "nn9536_4373_adam1"
  )
}
\dontshow{\}) # examplesIf}
}
