% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_databricks.R
\name{log_remove_connector.ConnectorDatabricksTable}
\alias{log_remove_connector.ConnectorDatabricksTable}
\title{Log Remove Operation for Databricks dbi connector}
\usage{
log_remove_connector.ConnectorDatabricksTable(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The ConnectorDatabricksTable object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\description{
Implementation of the log_remove_connector function for the
ConnectorDatabricksTable class.
}
