% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_databricks.R
\name{read_fs_databricks}
\alias{read_fs_databricks}
\title{Read files based on the extension}
\usage{
read_fs_databricks(path, ..., client = DatabricksClient())
}
\arguments{
\item{path}{Path to the file}

\item{...}{Other parameters for read's functions}

\item{client}{a databricks client as returned from \code{DatabricksClient}}
}
\value{
the result of the reading function
}
\description{
The aim of this function is to identify the extension on the file to dispatch it.
}
\examples{
temp_csv <- tempfile("iris", fileext = ".csv")
databrics_volume <- "/Volumes/my_adam/tester"
databricks_file_csv <- file.path(databrics_volume, basename(temp_csv))
write_fs_databricks(iris, databricks_file_csv)

# Read the file from databricks
read_fs_databricks(databricks_file_csv)
}
