% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{read_table_timepoint}
\alias{read_table_timepoint}
\title{Read Databricks table into dataframe}
\usage{
read_table_timepoint(connector_object, name, timepoint = NULL, version = NULL)
}
\arguments{
\item{connector_object}{\link{ConnectorDatabricksTable} object for interacting
with Databricks}

\item{name}{Table name}

\item{timepoint}{Timepoint in \href{https://docs.databricks.com/gcp/en/delta/history#delta-time-travel-syntax}{Delta time travel syntax}  # nolint
format.}

\item{version}{Table version generated by the operation.}
}
\value{
Data frame with values from the Databricks table
}
\description{
This function is used to read tables in a Databricks catalog with the option
of defining timepoint.
}
\details{
More info \href{https://docs.databricks.com/gcp/en/delta/history#what-is-delta-lake-time-travel}{here} # nolint
}
\examples{
\dontrun{
connector_object |>
    read_table_timepoint(name = "table_name",
                         timepoint = "2024-04-10 10:10:07 CEST"
     )
}
}
