% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_methods.R
\name{upload_directory_cnt.ConnectorDatabricksVolume}
\alias{upload_directory_cnt.ConnectorDatabricksVolume}
\title{Upload a directory}
\usage{
upload_directory_cnt.ConnectorDatabricksVolume(
  connector_object,
  dir,
  name = basename(dir),
  overwrite = zephyr::get_option("overwrite", "connector"),
  open = FALSE,
  ...
)
}
\arguments{
\item{...}{\link{ConnectorDatabricksVolume}: Additional parameters to pass to
the \code{\link[brickster:db_volume_dir_create]{brickster::db_volume_dir_create()}} method}
}
\value{
\link{ConnectorDatabricksVolume} object
}
\description{
Upload a directory
}
