% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/dbi_methods.R,
%   R/volume_methods.R
\name{write_cnt}
\alias{write_cnt}
\alias{write_cnt.connector_databricks_dbi}
\alias{write_cnt.ConnectorDatabricksVolume}
\title{Write content to the connector}
\usage{
write_cnt(connector_object, x, name, ...)

\method{write_cnt}{connector_databricks_dbi}(connector_object, x, name, ...)

\method{write_cnt}{ConnectorDatabricksVolume}(connector_object, x, name, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{x}{The object to write to the connection}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Addition write methods for databricks connectors implemented for \code{\link[connector:write_cnt]{connector::write_cnt()}}:

\itemize{
\item \link{connector_databricks_dbi}: Reuses the \code{\link[connector:read_cnt]{connector::read_cnt()}} method for \link[connector:connector_dbi]{connector::connector_dbi},
but always sets the \code{catalog} and \code{schema} as defined in when initializing the connector.
}

\itemize{
\item \link{ConnectorDatabricksVolume}: Reuses the \code{\link[connector:write_cnt]{connector::write_cnt()}} method for
\link{ConnectorDatabricksVolume}, but always sets the \code{catalog} and
\code{schema} as defined in when initializing the connector.
}
}
