% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{read_fs_databricks}
\alias{read_fs_databricks}
\alias{write_fs_databricks}
\title{Read files based on the extension}
\usage{
read_fs_databricks(path, client = DatabricksClient())

write_fs_databricks(
  x,
  file,
  ...,
  overwrite = TRUE,
  client = DatabricksClient()
)
}
\arguments{
\item{path}{Path to the file}

\item{client}{a databricks client as returned from \code{DatabricksClient}}

\item{x}{Object to write}

\item{file}{Path to write the file}

\item{...}{Other parameters for write functions}

\item{overwrite}{\code{logical} should the object be overwritten if it already exists}
}
\value{
the result of the reading function

The object \code{x} is returned invisibly
}
\description{
The aim of this function is to identify the extension on the file to dispatch it.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
temp_csv <- tempfile("iris", fileext = ".csv")
databrics_volume <- "/Volumes/my_adam/tester"
databricks_file_csv <- file.path(databrics_volume, basename(temp_csv))
try(write_fs_databricks(iris, databricks_file_csv))

# Read the file from databricks
try(read_fs_databricks(databricks_file_csv))
\dontshow{\}) # examplesIf}
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
temp_csv <- tempfile("iris", fileext = ".csv")
databrics_volume <- "/Volumes/my_adam/tester"
databricks_file_csv <- file.path(databrics_volume, basename(temp_csv))
try(write_fs_databricks(iris, databricks_file_csv))
\dontshow{\}) # examplesIf}
}
