% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{write_table_volume}
\alias{write_table_volume}
\title{Write data to a table using Databricks Volume}
\usage{
write_table_volume(
  connector_object,
  x,
  name,
  overwrite = zephyr::get_option("overwrite_write_table", "connector.databricks"),
  tags = NULL
)
}
\arguments{
\item{connector_object}{A \link{ConnectorDatabricksTable} object for interacting
with Databricks}

\item{x}{The data to be written to the table}

\item{name}{The name of the table}

\item{overwrite}{needs describtion}

\item{tags}{Named list containing tag names and tag values, e.g.
list("tag_name1" = "tag_value1", "tag_name2" = "tag_value2").
More info \href{https://docs.databricks.com/aws/en/database-objects/tags}{here}}
}
\value{
None
}
\description{
This function first writes parquet file to a temporary Databricks Volume and
then converts it to a table.
}
\examples{
\dontrun{
 write_table_volume(connector_object,
    data,
    "my_table",
    overwrite = zephyr::get_option("overwrite", "connector.databricks"),
    tags = list("tag_name1" = "tag_value1")
 )
}
}
