% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R
\name{cnt_disconnect}
\alias{cnt_disconnect}
\title{Disconnect (close) the connection of the connector}
\usage{
cnt_disconnect(connector_object, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Generic implementing of how to disconnect from the relevant connections.
Mostly relevant for DBI connectors.

\itemize{
\item \link[connector]{connector_dbi}: Uses \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}} to create a table reference to close a DBI connection.
}
}
