% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.databricks-package.R
\docType{package}
\name{connector.databricks-package}
\alias{connector.databricks}
\alias{connector.databricks-package}
\title{connector.databricks: Expand 'connector' Package for 'Databricks' Tables and Volumes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Expands the 'connector' \url{https://github.com/NovoNordisk-OpenSource/connector} package and provides a convenient interface for accessing and interacting with 'Databricks' \url{https://www.databricks.com} volumes and tables directly from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://novonordisk-opensource.github.io/connector.databricks/}
  \item \url{https://github.com/NovoNordisk-OpenSource/connector.databricks}
  \item Report bugs at \url{https://github.com/NovoNordisk-OpenSource/connector.databricks/issues}
}

}
\author{
\strong{Maintainer}: Vladimir Obucina \email{vlob@novonordisk.com}

Authors:
\itemize{
  \item Steffen Falgreen Larsen \email{sffl@novonordisk.com}
  \item Aksel Thomsen \email{oath@novonordisk.com}
  \item Cervan Girard \email{cgid@novonordisk.com}
}

Other contributors:
\itemize{
  \item Oliver Lundsgaard \email{ovlr@novonordisk.com} [contributor]
  \item Skander Mulder \email{sktf@novonordisk.com} [contributor]
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
