% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/volume_methods.R
\name{create_directory_cnt}
\alias{create_directory_cnt}
\alias{create_directory_cnt.ConnectorDatabricksVolume}
\title{Create a directory}
\usage{
create_directory_cnt(connector_object, name, open = TRUE, ...)

\method{create_directory_cnt}{ConnectorDatabricksVolume}(connector_object, name, open = TRUE, ...)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{name}{\link{character} The name of the directory to create}

\item{open}{create a new connector object}

\item{...}{\link{ConnectorDatabricksVolume}: Additional parameters to pass to the
\link[brickster:db_volume_dir_create]{brickster::db_volume_dir_create} method}
}
\value{
\link{invisible} connector_object.
}
\description{
Additional list content methods for Databricks connectors implemented for
\code{\link[connector:create_directory_cnt]{connector::create_directory_cnt()}}:

\itemize{
\item \link{ConnectorDatabricksVolume}: Reuses the \code{\link[connector:create_directory_cnt]{connector::create_directory_cnt()}}
method for \link{ConnectorDatabricksVolume}, but always
sets the \code{catalog}, \code{schema} and \code{path} as defined in when initializing the
connector.
}
}
