% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/volume_methods.R
\name{create_directory_cnt}
\alias{create_directory_cnt}
\alias{create_directory_cnt.ConnectorDatabricksVolume}
\title{Create a directory}
\usage{
create_directory_cnt(connector_object, name, ...)

\method{create_directory_cnt}{ConnectorDatabricksVolume}(connector_object, name, ..., open = TRUE)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{name}{\link{character} The name of the directory to create}

\item{...}{\link{ConnectorDatabricksVolume}: Additional parameters to pass to the \link[brickster:db_volume_dir_create]{brickster::db_volume_dir_create} method}

\item{open}{create a new connector object}
}
\value{
\link{ConnectorDatabricksVolume} object or \link{ConnectorDatabricksVolume} object of a newly built directory
}
\description{
Addition list content methods for databricks connectors implemented for \code{\link[connector:create_directory_cnt]{connector::create_directory_cnt()}}:
}
