% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_methods_utils.R
\name{download_cnt}
\alias{download_cnt}
\title{Download content from Databricks volume}
\usage{
download_cnt(connector_object, name, dest, ...)
}
\arguments{
\item{connector_object}{The \link{ConnectorDatabricksVolume} object}

\item{name}{The name of the file to download}

\item{dest}{The destination of the file}

\item{...}{Additional parameters to pass to the download method}
}
\value{
Response from the request
}
\description{
It use the \code{\link[=files_download_file]{files_download_file()}} method to download a file
from a Databricks volume
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# databricks_volume is a ConnectorDatabricksVolume object
databricks_volume \%>\%
  download_cnt("file.csv", "file.csv")

# This function is used by the method download_content
databricks_volume$download_content("file.csv", "file.csv")
\dontshow{\}) # examplesIf}
}
