% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/volume_methods.R
\name{download_directory_cnt}
\alias{download_directory_cnt}
\alias{download_directory_cnt.ConnectorDatabricksVolume}
\title{Download a directory}
\usage{
download_directory_cnt(connector_object, name, dir = name, ...)

\method{download_directory_cnt}{ConnectorDatabricksVolume}(connector_object, name, dir = basename(name), ...)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{name}{\link{character} The name of the directory to download}

\item{dir}{\link{character} Path to the directory to download to}

\item{...}{\link{ConnectorDatabricksVolume}: Additional parameters to pass to
the \code{\link[brickster:db_volume_dir_create]{brickster::db_volume_dir_create()}} method}
}
\value{
\link{invisible} connector_object.
}
\description{
Additional list content methods for Databricks connectors implemented for
\code{\link[connector:download_directory_cnt]{connector::download_directory_cnt()}}:

\itemize{
\item \link{ConnectorDatabricksVolume}: Reuses the
\code{\link[connector:download_directory_cnt]{connector::download_directory_cnt()}} method for
\link{ConnectorDatabricksVolume}, but always sets the
\code{catalog}, \code{schema} and \code{path} as defined in when initializing the connector.
}
}
