% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_methods.R
\name{download_directory_cnt.ConnectorDatabricksVolume}
\alias{download_directory_cnt.ConnectorDatabricksVolume}
\title{Download a directory}
\usage{
download_directory_cnt.ConnectorDatabricksVolume(
  connector_object,
  name,
  dir = basename(name),
  ...
)
}
\arguments{
\item{...}{\link{ConnectorDatabricksVolume}: Additional parameters to pass to
the \code{\link[brickster:db_volume_dir_create]{brickster::db_volume_dir_create()}} method}
}
\value{
\link{ConnectorDatabricksVolume} object
}
\description{
Download a directory
}
