% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{find_file_databricks}
\alias{find_file_databricks}
\title{Find File}
\usage{
find_file_databricks(name, root, client = DatabricksClient())
}
\arguments{
\item{name}{Name of a file}

\item{root}{Path to the root folder}

\item{client}{DatabricksClient}
}
\value{
A full name path to the file or a error if multiples files or 0.
}
\description{
Find File
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
iris_csv <- "iris.csv"
databrics_volume <- "/Volumes/amace_cdr_bronze_dev/nn9536_4373_adam/tester"
databricks_file_csv <- file.path(databrics_volume, iris_csv)
write_fs_databricks(iris, databricks_file_csv)
connector.databricks:::find_file_databricks("iris", root = databrics_volume)
\dontshow{\}) # examplesIf}
}
