% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_databricks.R, R/connector_imports.R
\name{log_remove_connector.ConnectorDatabricksTable}
\alias{log_remove_connector.ConnectorDatabricksTable}
\alias{log_remove_connector.ConnectorDatabricksVolume}
\alias{log_remove_connector}
\title{Log Remove Connector}
\usage{
\method{log_remove_connector}{ConnectorDatabricksTable}(connector_object, name, ...)

\method{log_remove_connector}{ConnectorDatabricksVolume}(connector_object, name, ...)

log_remove_connector(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The connector object to log the removal of.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters passed to the specific method implementation}
}
\value{
The result of the specific method implementation.
}
\description{
\itemize{
\item \link{ConnectorDatabricksTable}: Implementation of the \code{log_remove_connector}
function for the ConnectorDatabricksTable class.
}

\itemize{
\item \link{ConnectorDatabricksVolume}: Implementation of the \code{log_remove_connector}
function for the ConnectorDatabricksVolume class.
}

Additional log remove methods for Databricks connectors implemented for
\code{\link[connector:log_remove_connector]{connector::log_remove_connector()}}:
}
