% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_methods_utils.R
\name{remove_directory_cnt}
\alias{remove_directory_cnt}
\title{Remove a directory}
\usage{
remove_directory_cnt(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The \link{ConnectorDatabricksVolume} object}

\item{name}{The name of the directory to create}

\item{...}{Additional parameters to pass to the files_create_directory method}
}
\description{
Remove a directory
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# databricks_volume is a ConnectorDatabricksVolume object
databricks_volume \%>\%
  cnt_create_directory("folder")

# This function is used by the method create_directory
databricks_volume$create_directory("folder")
\dontshow{\}) # examplesIf}
}
