% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/volume_methods.R
\name{upload_directory_cnt}
\alias{upload_directory_cnt}
\alias{upload_directory_cnt.ConnectorDatabricksVolume}
\title{Upload a directory}
\usage{
upload_directory_cnt(
  connector_object,
  dir,
  name,
  overwrite = zephyr::get_option("overwrite", "connector"),
  open = FALSE,
  ...
)

\method{upload_directory_cnt}{ConnectorDatabricksVolume}(
  connector_object,
  dir,
  name = basename(dir),
  overwrite = zephyr::get_option("overwrite", "connector"),
  open = FALSE,
  ...
)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{dir}{\link{character} Path to the directory to upload}

\item{name}{\link{character} The name of the new directory to place the content in}

\item{overwrite}{Overwrite existing content if it exists in the connector?. Default: \code{FALSE}.}

\item{open}{\link{logical} Open the directory as a new connector object.}

\item{...}{\link{ConnectorDatabricksVolume}: Additional parameters to pass to
the \code{\link[brickster:db_volume_dir_create]{brickster::db_volume_dir_create()}} method}
}
\value{
\link{ConnectorDatabricksVolume} object
}
\description{
Addition list content methods for databricks connectors implemented for
\code{\link[connector:upload_directory_cnt]{connector::upload_directory_cnt()}}:
}
