% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fs_databricks.R
\name{write_fs_databricks}
\alias{write_fs_databricks}
\title{Write a file based on this extension}
\usage{
write_fs_databricks(
  x,
  file,
  ...,
  overwrite = TRUE,
  client = DatabricksClient()
)
}
\arguments{
\item{x}{Object to write}

\item{file}{Path to write the file}

\item{...}{Other parameters for write functions}

\item{overwrite}{\code{logical} should the object be overwritten if it already exists}

\item{client}{a databricks client as returned from \code{DatabricksClient}}
}
\value{
The object \code{x} is returned invisibly
}
\description{
Write a file based on this extension
}
\examples{
temp_csv <- tempfile("iris", fileext = ".csv")
databrics_volume <- "/Volumes/my_adam/tester"
databricks_file_csv <- file.path(databrics_volume, basename(temp_csv))
try(write_fs_databricks(iris, databricks_file_csv))

}
